feat: parse nested pydantic models `fetchone` and `fetchall` + add shortcuts for insert_query and update_query into `Database` (#2714)
* feat: add shortcuts for insert_query and update_query into `Database`
example: await db.insert("table_name", base_model)
* remove where from argument
* chore: code clean-up
* extension manager
* lnbits-qrcode  components
* parse date from dict
* refactor: make `settings` a fixture
* chore: remove verbose key names
* fix: time column
* fix: cast balance to `int`
* extension toggle vue3
* vue3 @input migration
* fix: payment extra and payment hash
* fix dynamic fields and ext db migration
* remove shadow on cards in dark theme
* screwed up and made more css pushes to this branch
* attempt to make chip component in settings dynamic fields
* dynamic chips
* qrscanner
* clean init admin settings
* make get_user better
* add dbversion model
* remove update_payment_status/extra/details
* traces for value and assertion errors
* refactor services
* add PaymentFiatAmount
* return Payment on api endpoints
* rename to get_user_from_account
* refactor: just refactor (#2740)
* rc5
* Fix db cache (#2741)
* [refactor] split services.py (#2742)
* refactor: spit `core.py` (#2743)
* refactor: make QR more customizable
* fix: print.html
* fix: qrcode options
* fix: white shadow on dark theme
* fix: datetime wasnt parsed in dict_to_model
* add timezone for conversion
* only parse timestamp for sqlite, postgres does it
* log internal payment success
* fix: export wallet to phone QR
* Adding a customisable border theme, like gradient (#2746)
* fixed mobile scan btn
* fix test websocket
* fix get_payments tests
* dict_to_model skip none values
* preimage none instead of defaulting to 0000...
* fixup test real invoice tests
* fixed pheonixd for wss
* fix nodemanager test settings
* fix lnbits funding
* only insert extension when they dont exist

---------

+++ lnbits/core/extensions/models.py
@@ -109,8 +109,8 @@ class ReleasePaymentInfo(BaseModel):
 
 
 class PayToEnableInfo(BaseModel):
-    required: Optional[bool] = False
-    amount: Optional[int] = None
+    amount: int
+    required: bool = False
     wallet: Optional[str] = None
 
 
@@ -120,6 +120,7 @@ class UserExtensionInfo(BaseModel):
 
 
 class UserExtension(BaseModel):
+    user: str
     extension: str
     active: bool
     extra: Optional[UserExtensionInfo] = None
@@ -372,29 +373,37 @@ class ExtensionRelease(BaseModel):
             return None
 
 
+class ExtensionMeta(BaseModel):
+    installed_release: Optional[ExtensionRelease] = None
+    latest_release: Optional[ExtensionRelease] = None
+    pay_to_enable: Optional[PayToEnableInfo] = None
+    payments: list[ReleasePaymentInfo] = []
+    dependencies: list[str] = []
+    archive: Optional[str] = None
+    featured: bool = False
+
+
 class InstallableExtension(BaseModel):
     id: str
     name: str
+    version: str
     active: Optional[bool] = False
     short_description: Optional[str] = None
     icon: Optional[str] = None
-    dependencies: list[str] = []
-    is_admin_only: bool = False
     stars: int = 0
-    featured = False
-    latest_release: Optional[ExtensionRelease] = None
-    installed_release: Optional[ExtensionRelease] = None
-    payments: list[ReleasePaymentInfo] = []
-    pay_to_enable: Optional[PayToEnableInfo] = None
-    archive: Optional[str] = None
+    meta: Optional[ExtensionMeta] = None
+
+    @property
+    def is_admin_only(self) -> bool:
+        return self.id in settings.lnbits_admin_extensions
 
     @property
     def hash(self) -> str:
-        if self.installed_release:
-            if self.installed_release.hash:
-                return self.installed_release.hash
+        if self.meta and self.meta.installed_release:
+            if self.meta.installed_release.hash:
+                return self.meta.installed_release.hash
             m = hashlib.sha256()
-            m.update(f"{self.installed_release.archive}".encode())
+            m.update(f"{self.meta.installed_release.archive}".encode())
             return m.hexdigest()
         return "not-installed"
 
@@ -432,15 +441,15 @@ class InstallableExtension(BaseModel):
 
     @property
     def installed_version(self) -> str:
-        if self.installed_release:
-            return self.installed_release.version
+        if self.meta and self.meta.installed_release:
+            return self.meta.installed_release.version
         return ""
 
     @property
     def requires_payment(self) -> bool:
-        if not self.pay_to_enable:
+        if not self.meta or not self.meta.pay_to_enable:
             return False
-        return self.pay_to_enable.required is True
+        return self.meta.pay_to_enable.required is True
 
     async def download_archive(self):
         logger.info(f"Downloading extension {self.name} ({self.installed_version}).")
@@ -448,12 +457,14 @@ class InstallableExtension(BaseModel):
         if ext_zip_file.is_file():
             os.remove(ext_zip_file)
         try:
-            assert self.installed_release, "installed_release is none."
+            assert (
+                self.meta and self.meta.installed_release
+            ), "installed_release is none."
 
             self._restore_payment_info()
 
             await asyncio.to_thread(
-                download_url, self.installed_release.archive_url, ext_zip_file
+                download_url, self.meta.installed_release.archive_url, ext_zip_file
             )
 
             self._remember_payment_info()
@@ -463,7 +474,11 @@ class InstallableExtension(BaseModel):
             raise AssertionError("Cannot fetch extension archive file") from exc
 
         archive_hash = file_hash(ext_zip_file)
-        if self.installed_release.hash and self.installed_release.hash != archive_hash:
+        if (
+            self.meta
+            and self.meta.installed_release.hash
+            and self.meta.installed_release.hash != archive_hash
+        ):
             # remove downloaded archive
             if ext_zip_file.is_file():
                 os.remove(ext_zip_file)
@@ -497,17 +512,18 @@ class InstallableExtension(BaseModel):
             self.short_description = config_json.get("short_description")
 
             if (
-                self.installed_release
-                and self.installed_release.is_github_release
+                self.meta
+                and self.meta.installed_release
+                and self.meta.installed_release.is_github_release
                 and config_json.get("tile")
             ):
                 self.icon = icon_to_github_url(
-                    self.installed_release.source_repo, config_json.get("tile")
+                    self.meta.installed_release.source_repo, config_json.get("tile")
                 )
 
         shutil.rmtree(self.ext_dir, True)
         shutil.copytree(Path(self.ext_upgrade_dir), Path(self.ext_dir))
-        logger.success(f"Extension {self.name} ({self.installed_version}) installed.")
+        logger.info(f"Extension {self.name} ({self.installed_version}) extracted.")
 
     def clean_extension_files(self):
         # remove downloaded archive
@@ -522,64 +538,54 @@ class InstallableExtension(BaseModel):
     def check_latest_version(self, release: Optional[ExtensionRelease]):
         if not release:
             return
-        if not self.latest_release:
-            self.latest_release = release
+        if not self.meta or not self.meta.latest_release:
+            meta = self.meta or ExtensionMeta()
+            meta.latest_release = release
+            self.meta = meta
             return
-        if version_parse(self.latest_release.version) < version_parse(release.version):
-            self.latest_release = release
+        if version_parse(self.meta.latest_release.version) < version_parse(
+            release.version
+        ):
+            self.meta.latest_release = release
 
     def find_existing_payment(
         self, pay_link: Optional[str]
     ) -> Optional[ReleasePaymentInfo]:
-        if not pay_link:
+        if not pay_link or not self.meta or not self.meta.payments:
             return None
         return next(
-            (p for p in self.payments if p.pay_link == pay_link),
+            (p for p in self.meta.payments if p.pay_link == pay_link),
             None,
         )
 
     def _restore_payment_info(self):
-        if not self.installed_release:
+        if (
+            not self.meta
+            or not self.meta.installed_release
+            or not self.meta.installed_release.pay_link
+            or not self.meta.installed_release.payment_hash
+        ):
             return
-        if not self.installed_release.pay_link:
-            return
-        if self.installed_release.payment_hash:
-            return
-        payment_info = self.find_existing_payment(self.installed_release.pay_link)
+        payment_info = self.find_existing_payment(self.meta.installed_release.pay_link)
         if payment_info:
-            self.installed_release.payment_hash = payment_info.payment_hash
+            self.meta.installed_release.payment_hash = payment_info.payment_hash
 
     def _remember_payment_info(self):
-        if not self.installed_release or not self.installed_release.pay_link:
+        if (
+            not self.meta
+            or not self.meta.installed_release
+            or not self.meta.installed_release.pay_link
+        ):
             return
         payment_info = ReleasePaymentInfo(
-            amount=self.installed_release.cost_sats,
-            pay_link=self.installed_release.pay_link,
-            payment_hash=self.installed_release.payment_hash,
+            amount=self.meta.installed_release.cost_sats,
+            pay_link=self.meta.installed_release.pay_link,
+            payment_hash=self.meta.installed_release.payment_hash,
         )
-        self.payments = [
-            p for p in self.payments if p.pay_link != payment_info.pay_link
+        self.meta.payments = [
+            p for p in self.meta.payments if p.pay_link != payment_info.pay_link
         ]
-        self.payments.append(payment_info)
-
-    @classmethod
-    def from_row(cls, data: dict) -> InstallableExtension:
-        meta = json.loads(data["meta"])
-        ext = InstallableExtension(**data)
-        if "installed_release" in meta:
-            ext.installed_release = ExtensionRelease(**meta["installed_release"])
-        if meta.get("pay_to_enable"):
-            ext.pay_to_enable = PayToEnableInfo(**meta["pay_to_enable"])
-        if meta.get("payments"):
-            ext.payments = [ReleasePaymentInfo(**p) for p in meta["payments"]]
-
-        return ext
-
-    @classmethod
-    def from_rows(cls, rows: Optional[list[Any]] = None) -> list[InstallableExtension]:
-        if rows is None:
-            rows = []
-        return [InstallableExtension.from_row(row) for row in rows]
+        self.meta.payments.append(payment_info)
 
     @classmethod
     async def from_github_release(
@@ -593,14 +599,17 @@ class InstallableExtension(BaseModel):
             return InstallableExtension(
                 id=github_release.id,
                 name=config.name,
+                version=latest_release.tag_name,
                 short_description=config.short_description,
                 stars=int(repo.stargazers_count),
                 icon=icon_to_github_url(
                     source_repo,
                     config.tile,
                 ),
-                latest_release=ExtensionRelease.from_github_release(
-                    source_repo, latest_release
+                meta=ExtensionMeta(
+                    latest_release=ExtensionRelease.from_github_release(
+                        source_repo, latest_release
+                    ),
                 ),
             )
         except Exception as e:
@@ -609,13 +618,14 @@ class InstallableExtension(BaseModel):
 
     @classmethod
     def from_explicit_release(cls, e: ExplicitRelease) -> InstallableExtension:
+        meta = ExtensionMeta(archive=e.archive, dependencies=e.dependencies)
         return InstallableExtension(
             id=e.id,
             name=e.name,
-            archive=e.archive,
+            version=e.version,
             short_description=e.short_description,
             icon=e.icon,
-            dependencies=e.dependencies,
+            meta=meta,
         )
 
     @classmethod
@@ -636,11 +646,13 @@ class InstallableExtension(BaseModel):
                     existing_ext = next(
                         (ee for ee in extension_list if ee.id == r.id), None
                     )
-                    if existing_ext:
-                        existing_ext.check_latest_version(ext.latest_release)
+                    if existing_ext and ext.meta:
+                        existing_ext.check_latest_version(ext.meta.latest_release)
                         continue
 
-                    ext.featured = ext.id in manifest.featured
+                    meta = ext.meta or ExtensionMeta()
+                    meta.featured = ext.id in manifest.featured
+                    ext.meta = meta
                     extension_list += [ext]
                     extension_id_list += [ext.id]
 
@@ -654,7 +666,9 @@ class InstallableExtension(BaseModel):
                         continue
                     ext = InstallableExtension.from_explicit_release(e)
                     ext.check_latest_version(release)
-                    ext.featured = ext.id in manifest.featured
+                    meta = ext.meta or ExtensionMeta()
+                    meta.featured = ext.id in manifest.featured
+                    ext.meta = meta
                     extension_list += [ext]
                     extension_id_list += [e.id]
             except Exception as e:
